[Refactor] Use Signature Algorithm instead of Authentication algorithm when creating and validating JWTs, use EdDSA(Ed25519) instead of HS512
