+++ src/main/java/com/dockerino/demo/api/AccountApi.java
@@ -1,10 +1,11 @@
-package com.dockerino.demo.security;
+package com.dockerino.demo.api;
 
 import com.dockerino.demo.model.User;
 import com.dockerino.demo.model.dtos.AuthResponse;
 import com.dockerino.demo.model.dtos.LoginRequest;
 import com.dockerino.demo.model.dtos.RegisterRequest;
 import com.dockerino.demo.model.dtos.UserInfo;
+import com.dockerino.demo.service.AuthService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
