+++ src/main/java/com/dockerino/demo/service/AuthService.java
@@ -1,4 +1,4 @@
-package com.dockerino.demo.security;
+package com.dockerino.demo.service;
 
 import com.dockerino.demo.model.User;
 import com.dockerino.demo.model.dtos.AuthResponse;
@@ -6,6 +6,8 @@ import com.dockerino.demo.model.dtos.LoginRequest;
 import com.dockerino.demo.model.dtos.RegisterRequest;
 import com.dockerino.demo.model.dtos.UserInfo;
 import com.dockerino.demo.repository.UserRepository;
+import com.dockerino.demo.security.CustomUserDetails;
+import com.dockerino.demo.security.JwtTokenProvider;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
