+++ src/main/resources/application.yml
@@ -9,5 +9,6 @@ spring:
     password: ${POSTGRES_PASSWORD}
 app:
   jwt:
-    secret: ${JWT_SECRET_KEY}
+    public-key-b64: ${PUBLIC_KEY}
+    private-key-b64: ${PRIVATE_KEY}
     expiration-ms: 300000
